requirements.txt
streamlit>=1.28.0
groq>=0.31.0
python-dotenv>=1.0.0
llama-index-core>=0.10.0
pymupdf>=1.26.0
markdown>=3.8.0 